+++ packages/stylelint-less/src/index.js
@@ -1,7 +1,7 @@
-import { createPlugin } from 'stylelint';
+import stylelint from 'stylelint';
 import rules from './rules';
 import { namespace } from './utils';
 
 export default Object.keys(rules).map((ruleName) => {
-	return createPlugin(namespace(ruleName), rules[ruleName]);
+	return stylelint.createPlugin(namespace(ruleName), rules[ruleName]);
 });