+++ packages/stylelint-less/src/utils/index.js
@@ -3,3 +3,5 @@ export { default as isValidVariable } from './isValidVariable';
 export { default as isLessVariable } from './isLessVariable';
 export { default as isHexColor } from './isHexColor';
 export { default as isIgnoredFunction } from './isIgnoredFunction';
+export { default as isStandardSyntaxAtRule } from './isStandardSyntaxAtRule';
+export { default as isValidHex } from './isValidHex';