+++ packages/stylelint-config-recommended-less/CHANGELOG.md
@@ -1,5 +1,31 @@
 # stylelint-config-recommended-less
 
+## 3.0.1
+
+### Patch Changes
+
+- f10f008: Fix compatibility with Stylelint 16
+- Updated dependencies [f10f008]
+  - stylelint-less@3.0.1
+
+## 3.0.0
+
+### Major Changes
+
+- eafa8ea: Export ESM from packages
+
+### Minor Changes
+
+- 0ed0822: Update packages to stylelint 16
+
+### Patch Changes
+
+- 23251e1: Update dependencies to the latest version
+- Updated dependencies [23251e1]
+- Updated dependencies [0ed0822]
+- Updated dependencies [eafa8ea]
+  - stylelint-less@3.0.0
+
 ## 2.0.0
 
 ### Major Changes